web: refactor wizard so that handlers live with renderers
It's better to have some DAMP (Distanct And Meaningful Phrases) for each step's
imports than to try and avoid it by cramming all the handlers somewhere other
than with the renderer.

This code moves each step into the same folder as its renderer, provides access
to the context so the buttons can be recalculated on the fly as needed (necessary
for the Bindings -> One Binding loop), and renames stuff that was poorly named
the first time around.